# Precompute output buffers