--- a/engine/bench_test.go
+++ b/engine/bench_test.go
@@ -85,7 +85,7 @@ func BenchmarkSingleQuery(b *testing.B) {
 	end := start.Add(1 * time.Hour)
 	step := time.Second * 30
 
-	query := "sum by (pod) (http_requests_total)"
+	query := "sum(http_requests_total)"
 	b.ResetTimer()
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {