--- a/executionplan/hash_aggregate.go
+++ b/executionplan/hash_aggregate.go
@@ -6,6 +6,8 @@ import (
 	"math"
 	"sync"
 
+	"github.com/prometheus/prometheus/model/labels"
+
 	"github.com/fpetkovski/promql-engine/model"
 
 	"github.com/prometheus/prometheus/promql/parser"
@@ -20,42 +22,89 @@ type aggregate struct {
 	by          bool
 	labels      []string
 	aggregation parser.ItemType
-	tables      []*aggregateTable
+
+	once   sync.Once
+	tables []*aggregateTable
+	series []labels.Labels
 }
 
 func NewAggregate(points *model.VectorPool, downstream VectorOperator, aggregation parser.ItemType, by bool, labels []string) (VectorOperator, error) {
-	keys := make([]groupingKey, 300000)
-	for i := 0; i < len(keys); i++ {
-		keys[i] = groupingKey{
-			once: &sync.Once{},
-		}
-	}
-	tables := make([]*aggregateTable, 30)
-	for i := 0; i < 30; i++ {
-		hashBuf := make([]byte, 128)
-		tables[i] = newAggregateTable(
-			newGroupingKeyGenerator(labels, !by, hashBuf),
-			func() *accumulator {
-				f, err := newAccumulator(aggregation)
-				if err != nil {
-					panic(err)
-				}
-				return f
-			},
-			keys,
-		)
-	}
-
 	return &aggregate{
 		downstream: downstream,
-		tables:     tables,
-
 		vectorPool: points,
 
 		by:          by,
 		aggregation: aggregation,
 		labels:      labels,
 	}, nil
+}
+
+func (a *aggregate) Series(ctx context.Context) ([]labels.Labels, error) {
+	var err error
+	a.once.Do(func() { err = a.initOutputBuffers(ctx) })
+	if err != nil {
+		return nil, err
+	}
+
+	return a.series, nil
+}
+
+func (a *aggregate) initOutputBuffers(ctx context.Context) error {
+	series, err := a.downstream.Series(ctx)
+	if err != nil {
+		return err
+	}
+
+	inputCache := make([]uint64, len(series))
+	outputMap := make(map[uint64]*aggregateResult)
+	outputCache := make([]*aggregateResult, 0)
+
+	for i := 0; i < len(series); i++ {
+		buf := make([]byte, 128)
+		hash, _, lbls := hashMetric(series[i], !a.by, a.labels, buf)
+
+		output, ok := outputMap[hash]
+		if !ok {
+			output = &aggregateResult{
+				metric:   lbls,
+				sampleID: uint64(len(outputCache)),
+			}
+			outputMap[hash] = output
+			outputCache = append(outputCache, output)
+		}
+
+		inputCache[i] = output.sampleID
+	}
+
+	var wg sync.WaitGroup
+	a.series = make([]labels.Labels, len(outputCache))
+	for i := 0; i < len(outputCache); i++ {
+		wg.Add(1)
+		go func(i int) {
+			defer wg.Done()
+			a.series[i] = outputCache[i].metric
+		}(i)
+	}
+	wg.Wait()
+
+	tables := make([]*aggregateTable, 10)
+	for i := 0; i < len(tables); i++ {
+		wg.Add(1)
+		go func(i int) {
+			defer wg.Done()
+			tables[i] = newAggregateTable(inputCache, outputCache, func() *accumulator {
+				f, err := newAccumulator(a.aggregation)
+				if err != nil {
+					panic(err)
+				}
+				return f
+			})
+		}(i)
+	}
+	wg.Wait()
+
+	a.tables = tables
+	return nil
 }
 
 func (a *aggregate) GetPool() *model.VectorPool {
@@ -71,6 +120,11 @@ func (a *aggregate) Next(ctx context.Context) ([]model.StepVector, error) {
 		return nil, nil
 	}
 	defer a.downstream.GetPool().PutVectors(in)
+
+	a.once.Do(func() { err = a.initOutputBuffers(ctx) })
+	if err != nil {
+		return nil, err
+	}
 
 	result := make([]model.StepVector, len(in))
 	var wg sync.WaitGroup
@@ -98,6 +152,7 @@ type accumulator struct {
 	AddFunc   func(v float64)
 	ValueFunc func() float64
 	HasValue  func() bool
+	Reset     func()
 }
 
 func newAccumulator(expr parser.ItemType) (*accumulator, error) {
@@ -113,6 +168,10 @@ func newAccumulator(expr parser.ItemType) (*accumulator, error) {
 			},
 			ValueFunc: func() float64 { return value },
 			HasValue:  func() bool { return hasValue },
+			Reset: func() {
+				hasValue = false
+				value = 0
+			},
 		}, nil
 	case "max":
 		var value float64
@@ -123,6 +182,10 @@ func newAccumulator(expr parser.ItemType) (*accumulator, error) {
 			},
 			ValueFunc: func() float64 { return value },
 			HasValue:  func() bool { return hasValue },
+			Reset: func() {
+				hasValue = false
+				value = 0
+			},
 		}, nil
 	case "min":
 		var value float64
@@ -133,6 +196,10 @@ func newAccumulator(expr parser.ItemType) (*accumulator, error) {
 			},
 			ValueFunc: func() float64 { return value },
 			HasValue:  func() bool { return hasValue },
+			Reset: func() {
+				hasValue = false
+				value = 0
+			},
 		}, nil
 	case "count":
 		var value float64
@@ -143,6 +210,10 @@ func newAccumulator(expr parser.ItemType) (*accumulator, error) {
 			},
 			ValueFunc: func() float64 { return value },
 			HasValue:  func() bool { return hasValue },
+			Reset: func() {
+				hasValue = false
+				value = 0
+			},
 		}, nil
 	case "avg":
 		var count, sum float64
@@ -154,6 +225,11 @@ func newAccumulator(expr parser.ItemType) (*accumulator, error) {
 			},
 			ValueFunc: func() float64 { return sum / count },
 			HasValue:  func() bool { return hasValue },
+			Reset: func() {
+				hasValue = false
+				sum = 0
+				count = 0
+			},
 		}, nil
 	}
 	return nil, fmt.Errorf("unknown aggregation function %s", t)