This change allows each operator to precompute its input and output buffers by calling `Series()` on its downstream operators.

Even though precomputing buffers is additional work at the start of the execution, it allows us to:
* allocate memory very precisely
* use arrays instead of maps for lookups
* use very tight loops with no conditional statements

Benchmark before:
```
goos: darwin
goarch: arm64
pkg: github.com/fpetkovski/promql-engine/engine
BenchmarkOldEngine
BenchmarkOldEngine/vector_selector
BenchmarkOldEngine/vector_selector/current_engine
BenchmarkOldEngine/vector_selector/current_engine-8         	      43	  26262838 ns/op	16147712 B/op	   69122 allocs/op
BenchmarkOldEngine/vector_selector/new_engine
BenchmarkOldEngine/vector_selector/new_engine-8             	      36	  33871655 ns/op	37037675 B/op	  104161 allocs/op
BenchmarkOldEngine/aggregation
BenchmarkOldEngine/aggregation/current_engine
BenchmarkOldEngine/aggregation/current_engine-8             	      22	  46908994 ns/op	 4932668 B/op	   73845 allocs/op
BenchmarkOldEngine/aggregation/new_engine
BenchmarkOldEngine/aggregation/new_engine-8                 	      62	  18749468 ns/op	31787828 B/op	  377510 allocs/op
BenchmarkOldEngine/aggregation_by_pod
BenchmarkOldEngine/aggregation_by_pod/current_engine
BenchmarkOldEngine/aggregation_by_pod/current_engine-8      	       8	 130925594 ns/op	87268860 B/op	  568656 allocs/op
BenchmarkOldEngine/aggregation_by_pod/new_engine
BenchmarkOldEngine/aggregation_by_pod/new_engine-8          	      33	  38140593 ns/op	78981717 B/op	 2085462 allocs/op
```

Benchmark after
```
goos: darwin
goarch: arm64
pkg: github.com/fpetkovski/promql-engine/engine
BenchmarkOldEngine
BenchmarkOldEngine/vector_selector
BenchmarkOldEngine/vector_selector/current_engine
BenchmarkOldEngine/vector_selector/current_engine-8         	      43	  26150992 ns/op	16147309 B/op	   69122 allocs/op
BenchmarkOldEngine/vector_selector/new_engine
BenchmarkOldEngine/vector_selector/new_engine-8             	      79	  15771956 ns/op	42044563 B/op	   78790 allocs/op
BenchmarkOldEngine/aggregation
BenchmarkOldEngine/aggregation/current_engine
BenchmarkOldEngine/aggregation/current_engine-8             	      24	  46801087 ns/op	 4931594 B/op	   73844 allocs/op
BenchmarkOldEngine/aggregation/new_engine
BenchmarkOldEngine/aggregation/new_engine-8                 	      94	  13088061 ns/op	26100849 B/op	   74025 allocs/op
BenchmarkOldEngine/aggregation_by_pod
BenchmarkOldEngine/aggregation_by_pod/current_engine
BenchmarkOldEngine/aggregation_by_pod/current_engine-8      	       8	 129737219 ns/op	87268186 B/op	  568653 allocs/op
BenchmarkOldEngine/aggregation_by_pod/new_engine
BenchmarkOldEngine/aggregation_by_pod/new_engine-8          	      76	  15833226 ns/op	38186500 B/op	  161411 allocs/op

```